files rearranged with docker utils
